+++ lib/honeybadger/util/sanitizer.rb
@@ -3,41 +3,44 @@ module Honeybadger
     class Sanitizer
       OBJECT_WHITELIST = [Hash, Array, String, Integer, Float, TrueClass, FalseClass, NilClass]
 
+      FILTERED_REPLACEMENT = '[FILTERED]'.freeze
+
       def initialize(opts = {})
-        @max_depth = opts[:max_depth] || 20
-        @filters = Array(opts[:filters])
+        @max_depth = opts.fetch(:max_depth, 20)
+        @filters = Array(opts.fetch(:filters, nil)).collect do |f|
+          f.kind_of?(Regexp) ? f : f.to_s
+        end
       end
 
-      # Removes non-serializable data and truncates to max depth.
       def sanitize(data, depth = 0, stack = [])
         return '[possible infinite recursion halted]' if stack.any?{|item| item == data.object_id }
 
-        if data.respond_to?(:to_hash)
+        if data.kind_of?(String)
+          sanitize_string(data)
+        elsif data.respond_to?(:to_hash)
           return '[max depth reached]' if depth >= max_depth
-          data.to_hash.reduce({}) do |result, (key, value)|
-            result.merge(key => sanitize(value, depth+1, stack + [data.object_id]))
+          hash = data.to_hash
+          new_hash = {}
+          new_stack = stack.dup.push(data.object_id)
+          hash.each_pair do |key, value|
+            k = key.kind_of?(Symbol) ? key : sanitize(key, depth+1, stack)
+            if filter_key?(k)
+              new_hash[k] = FILTERED_REPLACEMENT
+            else
+              new_hash[k] = sanitize(value, depth+1, new_stack)
+            end
           end
+          new_hash
         elsif data.respond_to?(:to_ary)
           return '[max depth reached]' if depth >= max_depth
+          new_stack = stack.dup.push(data.object_id)
           data.to_ary.collect do |value|
-            sanitize(value, depth+1, stack + [data.object_id])
+            sanitize(value, depth+1, new_stack)
           end
         elsif OBJECT_WHITELIST.any? {|c| data.kind_of?(c) }
           data
         else
-          data.to_s
-        end
-      end
-
-      def filter(hash)
-        {}.tap do |filtered_hash|
-          hash.each_pair do |key, value|
-            if value.respond_to?(:to_hash)
-              filtered_hash[key] = filter(hash[key])
-            else
-              filtered_hash[key] = filter_key?(key) ? '[FILTERED]' : hash[key]
-            end
-          end
+          sanitize_string(data.to_s)
         end
       end
 
@@ -53,6 +56,10 @@ module Honeybadger
 
       private
 
+      VALID_ENCODINGS = [Encoding::UTF_8, Encoding::ISO_8859_1].freeze
+      ENCODE_OPTS = { invalid: :replace, undef: :replace, replace: '?'.freeze }.freeze
+      UTF8_STRING = ''.freeze
+
       attr_reader :max_depth, :filters
 
       def filter_key?(key)
@@ -60,10 +67,27 @@ module Honeybadger
 
         filters.any? do |filter|
           if filter.is_a?(Regexp)
-            key.to_s =~ filter
+            filter =~ key.to_s
           else
             key.to_s.eql?(filter.to_s)
           end
+        end
+      end
+
+      def valid_encoding?(data)
+         data.valid_encoding? && (
+           VALID_ENCODINGS.include?(data.encoding) ||
+           VALID_ENCODINGS.include?(Encoding.compatible?(UTF8_STRING, data))
+         )
+      end
+
+      def sanitize_string(data)
+        return data if valid_encoding?(data)
+
+        if data.encoding == Encoding::UTF_8
+          data.encode(Encoding::UTF_16, ENCODE_OPTS).encode!(Encoding::UTF_8)
+        else
+          data.encode(Encoding::UTF_8, ENCODE_OPTS)
         end
       end
     end